feat: set static MariaDB values in model, refactor MariaDB model (#20)
